Move wikidata out of its own module
